pyproject.toml
[project]
name = "laboratory-11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"




[dependency-groups]
integration = [
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "boto3>=1.42.1",
    "clean-text>=0.6.0",
    "fastapi>=0.118.2",
    "joblib>=1.5.2",
    "pydantic>=2.12.0",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
]
deployment = [
    "onnxscript>=0.5.6",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "boto3>=1.42.1",
    "clean-text>=0.6.0",
    "joblib>=1.5.2",
    "pydantic>=2.12.0",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "pydantic-settings>=2.12.0",
    "skl2onnx>=1.19.1",
]
inference = [
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "fastapi>=0.118.2",
    "tokenizers>=0.22.1",
]

[tool.ruff]
line-length = 140


[tool.ruff.lint]
# Config found on: https://www.reddit.com/r/Python/comments/15o73kq/what_ruff_config_do_you_use/
# Read more here https://beta.ruff.rs/docs/rules/
# By default, Ruff enables Flake8's E and F rules
# Pyflakes - F, pycodestyle - E, W
# flake8-builtins - A
# Pylint - PLC, PLE, PLW
# isort - I
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I']

[tool.uv]
package = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["lab11_lib"]
exclude = ["tests*"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = ["src", "."]
testpaths = ["tests"]

 # I had a problem with running tests via pytest, when inside a test I import my own module
 # and I've got error "my_own_module" import not found
 # I've found couple of issues e.g: https://github.com/astral-sh/uv/issues/9291
 # and tips proposed in it helped me to resolve that issue
